Split remote bridge complexity
