+++ packages/net/gateway-stack/src/remote-bridge-util.ts
@@ -1,0 +1,3 @@
+export function isRecord(value: unknown): value is Readonly<Record<string, unknown>> {
+  return typeof value === "object" && value !== null;
+}
